refactor(chat): extract assistant message cards
